feat: add processed event (idempotent)
