+++ common-module/src/main/java/aop/IdempotentEventPayload.java
@@ -1,0 +1,7 @@
+package aop;
+
+public interface IdempotentEventPayload {
+    String eventId();
+    String topic();
+    String topicKey();
+}
