+++ order-event/src/main/java/com/leui/orderconsumer/domain/order/dto/OrderEvent.java
@@ -1,0 +1,11 @@
+package com.leui.orderconsumer.domain.order.dto;
+
+import aop.IdempotentEventPayload;
+import enumtype.OrderStatus;
+
+public record OrderEvent(
+        String eventId,
+        String topic,
+        String topicKey
+) implements IdempotentEventPayload {
+}
