+++ order-event/src/main/java/com/leui/orderconsumer/domain/order/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package com.leui.orderconsumer.domain.order.repository;
+
+import com.leui.orderconsumer.domain.order.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, String> {
+}
