--- order-event/src/main/java/com/leui/orderconsumer/repository/order/OrderRepository.java
@@ -1,6 +1,0 @@
-package com.leui.orderconsumer.repository.order;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, String> {
-}
