--- order-event/src/main/java/com/leui/orderconsumer/repository/prcessedevent/ProcessedEventRepository.java
@@ -1,6 +1,0 @@
-package com.leui.orderconsumer.repository.prcessedevent;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProcessedEventRepository extends JpaRepository<ProcessedEvent, String> {
-}
