+++ store-event/src/main/java/com/leui/storeevent/domain/deal/dto/DealEvent.java
@@ -1,0 +1,13 @@
+package com.leui.storeevent.domain.deal.dto;
+
+import aop.IdempotentEventPayload;
+
+public record DealEvent(
+        String eventId,
+        String topic,
+        String topicKey,
+        Long dealId,
+        Integer quantity,
+        String orderId
+) implements IdempotentEventPayload {
+}
