+++ store-event/src/main/java/com/leui/storeevent/domain/deal/dto/RemoveDealReservation.java
@@ -1,0 +1,11 @@
+package com.leui.storeevent.domain.deal.dto;
+
+import aop.IdempotentEventPayload;
+
+public record RemoveDealReservation(
+        String eventId,
+        String topic,
+        String topicKey,
+        String orderId
+) implements IdempotentEventPayload {
+}
