+++ store-event/src/main/java/com/leui/storeevent/domain/prcessedevent/aop/TransactionalIdempotentEvent.java
@@ -1,0 +1,11 @@
+package com.leui.storeevent.domain.prcessedevent.aop;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface TransactionalIdempotentEvent {
+}
