+++ store-event/src/main/java/com/leui/storeevent/domain/prcessedevent/repository/ProcessedEventRepository.java
@@ -1,0 +1,7 @@
+package com.leui.storeevent.domain.prcessedevent.repository;
+
+import com.leui.storeevent.domain.prcessedevent.entity.ProcessedEvent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProcessedEventRepository extends JpaRepository<ProcessedEvent, String> {
+}
