+++ user-event/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation 'org.springframework.kafka:spring-kafka'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-jdbc'
+    implementation 'org.springframework.boot:spring-boot-starter-aop'
     implementation project(':common-module')
 
     runtimeOnly 'org.postgresql:postgresql'
