+++ user-event/src/main/java/com/leui/userevent/domain/prcessedevent/repository/ProcessedEventRepository.java
@@ -1,0 +1,7 @@
+package com.leui.userevent.domain.prcessedevent.repository;
+
+import com.leui.userevent.domain.prcessedevent.entity.ProcessedEvent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProcessedEventRepository extends JpaRepository<ProcessedEvent, String> {
+}
