+++ user-event/src/main/java/com/leui/userevent/domain/user/dto/UserEvent.java
@@ -1,0 +1,15 @@
+package com.leui.userevent.domain.user.dto;
+
+import aop.IdempotentEventPayload;
+
+import java.math.BigDecimal;
+
+public record UserEvent(
+        String eventId,
+        String topic,
+        String topicKey,
+        Long userId,
+        BigDecimal totalAmount,
+        BigDecimal usedPoint
+) implements IdempotentEventPayload {
+}
