+++ user-event/src/main/java/com/leui/userevent/domain/user/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.leui.userevent.domain.user.repository;
+
+import com.leui.userevent.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
