--- user-event/src/main/java/com/leui/userevent/repository/User.java
@@ -1,49 +1,0 @@
-package com.leui.userevent.repository;
-
-import enumtype.Role;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Setter
-    @Column(nullable = false)
-    private String password;
-
-    private BigDecimal point;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Role role;
-
-    public User(String email, String password, Role role) {
-        this.email = email;
-        this.password = password;
-        this.role = role;
-    }
-
-    public void subtractPoint(BigDecimal used) {
-        this.point = this.point.subtract(used);
-    }
-
-    public void accumulatePoint(BigDecimal point) {
-        this.point = this.point.add(point);
-    }
-
-}
