--- user-event/src/main/java/com/leui/userevent/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.leui.userevent.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
